bot/requirements.txt
discord.py
wavelink
httpx